DermaGenieBackend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic[email]
bcrypt==4.0.1
passlib==1.7.4
python-jose
python-dotenv
cloudinary
python-multipart
alembic